Aresponse handler toegevoegd in utils - kijk even hoe ik deze implementeer in de vehicle fragment - kijk ook even hoe ik deze implementeer in de vehiclerepository
