+++ app/src/main/java/nl/otters/elbho/repositories/VehicleRepository.kt
@@ -3,25 +3,23 @@ package nl.otters.elbho.repositories
 import android.content.Context
 import android.util.Log
 import android.view.View
-import android.widget.TextView
-import android.widget.Toast
-import androidx.core.content.ContextCompat
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.google.android.material.snackbar.Snackbar
 import nl.otters.elbho.R
 import nl.otters.elbho.factories.RetrofitFactory
 import nl.otters.elbho.models.Vehicle
 import nl.otters.elbho.services.VehicleService
+import nl.otters.elbho.utils.ResponseHandler
 import nl.otters.elbho.utils.SharedPreferences
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
 
-class VehicleRepository(private val context: Context, private val currView: View) {
+class VehicleRepository(private val context: Context, private val curView: View) {
     private val vehicleService: VehicleService =
         RetrofitFactory.get().create(VehicleService::class.java)
+    private val responseHandler: ResponseHandler = ResponseHandler(context, curView)
 
     fun getAllVehicleReservations(options: Vehicle.CarReservationOptions?): LiveData<ArrayList<Vehicle.CarWithReservations>> {
         val vehicleReservations: MutableLiveData<ArrayList<Vehicle.CarWithReservations>> =
@@ -36,7 +34,7 @@ class VehicleRepository(private val context: Context, private val currView: View
                     if (response.code() == 200 && response.body() != null) {
                         vehicleReservations.value = response.body()!!
                     } else {
-                        errorMsg(R.string.error_api_vehicle)
+                        responseHandler.errorMessage(R.string.error_api_vehicle)
                     }
                 }
 
@@ -45,7 +43,7 @@ class VehicleRepository(private val context: Context, private val currView: View
                     t: Throwable
                 ) {
                     Log.e("HTTP Vehicles: ", "Could not fetch data", t)
-                    errorMsg(R.string.error_api_vehicle)
+                    responseHandler.errorMessage(R.string.error_api_vehicle)
                 }
             })
         return vehicleReservations
@@ -66,13 +64,13 @@ class VehicleRepository(private val context: Context, private val currView: View
                 if (response.code() == 200 && response.body() != null) {
                     vehicleReservations.value = response.body()!!
                 } else {
-                    errorMsg(R.string.error_api_vehicle)
+                    responseHandler.errorMessage(R.string.error_api_vehicle)
                 }
             }
 
             override fun onFailure(call: Call<ArrayList<Vehicle.Reservation>>, t: Throwable) {
                 Log.e("HTTP Vehicles: ", "Could not fetch data", t)
-                errorMsg(R.string.error_api_vehicle)
+                responseHandler.errorMessage(R.string.error_api_vehicle)
             }
         })
         return vehicleReservations
@@ -86,15 +84,15 @@ class VehicleRepository(private val context: Context, private val currView: View
                     response: Response<Unit>
                 ) {
                     if (response.code() == 201 && response.body() != null) {
-                        succesMsg(R.string.snackbar_vehicle_reserved)
+                        responseHandler.succesMessage(R.string.snackbar_vehicle_reserved)
                     } else {
-                        errorMsg(R.string.error_api_vehicle)
+                        responseHandler.errorMessage(R.string.error_api_vehicle)
                     }
                 }
 
                 override fun onFailure(call: Call<Unit>, t: Throwable) {
                     Log.e("HTTP Vehicles: ", "Could not fetch data", t)
-                    errorMsg(R.string.error_api_vehicle)
+                    responseHandler.errorMessage(R.string.error_api_vehicle)
                 }
             })
     }
@@ -107,50 +105,17 @@ class VehicleRepository(private val context: Context, private val currView: View
                     response: Response<Unit>
                 ) {
                     if (response.code() == 200 && response.body() != null) {
-                        succesMsg(R.string.snackbar_vehicle_cancelled)
+                        responseHandler.succesMessage(R.string.snackbar_vehicle_cancelled)
                     } else {
-                        errorMsg(R.string.error_api_vehicle)
+                        responseHandler.errorMessage(R.string.error_api_vehicle)
                     }
                 }
 
                 override fun onFailure(call: Call<Unit>, t: Throwable) {
                     Log.e("HTTP Vehicles: ", "Could not fetch data", t)
-                    errorMsg(R.string.error_api_vehicle)
+                    responseHandler.errorMessage(R.string.error_api_vehicle)
                 }
             })
-    }
-
-    private fun errorMsg(error_string: Int) {
-        Toast.makeText(
-            context,
-            error_string,
-            Toast.LENGTH_LONG
-        ).show()
-    }
-
-    private fun succesMsg(succes_string: Int) {
-        val snackbarDialog = Snackbar.make(
-            currView,
-            succes_string,
-            Snackbar.LENGTH_LONG
-        )
-        val snackbarView = snackbarDialog.view
-        snackbarView.setBackgroundColor(
-            ContextCompat.getColor(
-                context,
-                R.color.vehicle_snackBar_bg_col
-            )
-        )
-        val snackbarTextView =
-            snackbarView.findViewById<TextView>(R.id.snackbar_text)
-        snackbarTextView.setCompoundDrawablesRelativeWithIntrinsicBounds(
-            R.drawable.ic_check_circle_24dp,
-            0,
-            0,
-            0
-        )
-        snackbarTextView.compoundDrawablePadding = 75
-        snackbarDialog.show()
     }
 
     private fun getAuthToken(): String {
